added scale and rotate support

+++ source/FirstPersonNoClip.h
@@ -11,7 +11,7 @@ class FirstPersonNoClip
         ~FirstPersonNoClip();
         bool mouseMoved(const OIS::MouseEvent& evt);
         const Ogre::Quaternion& getOrientation();
-        void update(float timeSinceLastFrame, OIS::Keyboard* inputManager);
+        void update(double timeSinceLastFrame, OIS::Keyboard* inputManager);
         const Ogre::Vector3&    getPosition();
         Ogre::Ray getCameraToViewportRay(Ogre::Real screenx, Ogre::Real screeny);
     private:
